pyproject.toml
# -----------------------------
# Build system
# -----------------------------
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.hatch.build.targets.wheel]
packages = ["src/studio"]

# -----------------------------
# Dependency groups (PEP 735)
# -----------------------------
[dependency-groups]
dev = [
    "isort>=7.0.0",
    "mypy>=1.18.2",
    "nest-asyncio",
    "ruff>=0.14.5",
    "toml-sort>=0.24.3",
    "typing-extensions>=4.15.0",
]

# -----------------------------
# Project metadata + PyPI runtime deps
# -----------------------------
[project]
name = "MCP-demo"
description = "MCP demo library"
readme = "README.md"
license = { text = "NONE" }
authors = [
  { name = "RaoulG", email = "Raoul.Grouls@han.nl" },
  { name = "Hoog-V", email = "victor.hogeweij@han.nl" },
]
classifiers = [
  "Environment :: Console",
  "Framework :: FastAPI",
  "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.11"
dynamic = ["version"]

dependencies = [
    "fastmcp>=2.10.1",
    "fastapi",
    "httpx",
    "typedb-driver==2.29.2",
    "langfuse", 
    "anthropic>=0.42.0",
    "openai",
    "loguru>=0.7.3",
    "ollama>=0.4.5",
    "python-dotenv>=1.0.1",
    "qdrant-client>=1.15.1", 
    "sentence-transformers>=5.1.2", 
    "pyzotero>=1.7.3"
    , "uvicorn[standard]>=0.38.0,<0.39"]

# -----------------------------
# Testing
# -----------------------------
[tool.pytest.ini_options]
pythonpath = ["client", "server"]
addopts = ["--import-mode=importlib"]

# -----------------------------
# Pixi (Conda) — merged here for clarity
# -----------------------------
[tool.pixi.workspace]
channels = [
  "https://fast.prefix.dev/conda-forge",
  "https://repo.prefix.dev/pypdfium2-public",
  "https://repo.prefix.dev/pytorch",
]
platforms = ["linux-64", "osx-64", "osx-arm64"]

# Host deps (for builds like docling from PyPI)
[tool.pixi.host-dependencies]
compilers = ">=1.9.0"
pkg-config = ">=0.29.2"

# Conda-side runtime dependencies
[tool.pixi.dependencies]
python = "3.12.*"
python-dotenv = ">=1.1.0"
loguru = ">=0.7.3,<0.8"
lefthook = ">=1.11.13"
lychee = ">=0.15.1"
einops = ">=0.8.1"
ruff = ">=0.11.13"
isort = ">=6.0.1"
mypy = ">=1.16.0"
typing_extensions = ">=4.14.0"
nodejs = ">=22.13.0,<22.14"
itsdangerous = ">=2.2.0,<3"

[tool.pixi.tasks]
frontend_build="rm -r kg && cd src/frontend && npm install && npm run build && mv dist ../../kg"
backend_no_docker="BASE_URL=http://localhost:10090 PYTHONPATH=src python -m uvicorn src.backend.main:app --host 0.0.0.0 --port 10090  --reload"
backend="docker compose up"
mcp_server="PYTHONPATH=src python src/mcp_servers/main.py"
llm_worker="PYTHONPATH=src python src/llm_worker/main.py"
vscode = "code ."
lint = "ruff check"
typecheck = "mypy ."
test = "pytest -q"

# Activation env
[tool.pixi.activation.env]
PYTHONNOUSERSITE = "1"
LD_LIBRARY_PATH = "$CONDA_PREFIX/lib:$LD_LIBRARY_PATH"

# Platform-specific deps
[tool.pixi.target.linux-64.dependencies]
python-devtools = ">=0.12.2"         # Optional but useful
# cuda-toolkit = "12.9.0.*"   # For use with CUDA

# -----------------------------
# Mypy
# -----------------------------
[tool.mypy]
allow_redefinition = true
disable_error_code = [
    "call-arg",        # unexpected keyword arguments
    "import-untyped",  # missing stubs like requests
    "import-not-found",
    "call-overload",
    "arg-type",
    "misc",
]
pretty = true
exclude = [
    "^\\.pixi/",  # ignore the .pixi folder at repo root
    "^\\studio-project/"
]

# -----------------------------
# Ruff
# -----------------------------
[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "C4", "PT"]

[tool.ruff]
exclude = [".pixi", "tests", "studio-project"]
